fix(devtools): rename bridge
