+++ android/sdk/src/main/java/com/tencent/mtt/hippy/bridge/HippyBridgeImpl.java
@@ -299,7 +299,7 @@ public class HippyBridgeImpl implements HippyBridge, DevRemoteDebugProxy.OnRecei
     public native boolean runScriptFromUri(String uri, AssetManager assetManager,
             boolean canUseCodeCache, String codeCacheDir, long V8RuntimeId, NativeCallback callback);
 
-    public native void destroy(long runtimeId, boolean useLowMemoryMode, boolean is_reload, NativeCallback callback);
+    public native void destroy(long runtimeId, boolean useLowMemoryMode, boolean isReload, NativeCallback callback);
 
     public native void callFunction(String action, long runtimeId, NativeCallback callback,
             ByteBuffer buffer, int offset, int length);
