+++ android/sdk/src/main/jni/CMakeLists.txt
@@ -63,7 +63,7 @@ endif ()
 
 # region source set
 set(SOURCE_SET
-    src/bridge/bridge.cc
+    src/bridge/adr_bridge.cc
     src/bridge/entry.cc
     src/bridge/java2js.cc
     src/bridge/js2java.cc
