+++ android/sdk/src/main/jni/include/bridge/adr_bridge.h
