+++ core/include/core/core.h
@@ -47,6 +47,10 @@
 #ifdef JS_V8
 #include "core/napi/v8/js_native_api_v8.h"
 #include "core/napi/v8/js_native_turbo_v8.h"
+#include "core/inspector/bridge.h"
+#ifndef V8_WITHOUT_INSPECTOR
+#include "core/inspector/v8_inspector_client_impl.h"
+#endif
 #else
 #include "core/napi/jsc/js_native_api_jsc.h"
 #include "core/napi/jsc/js_native_jsc_helper.h"
